# Reduce allocations in TSI TagSets implementation