… tsdb/testdata/line-protocol-1M.txt.gz