@hercules-influx 2018-08-10
During a run of [megacheck](https://github.com/dominikh/go-tools) the following issues were discovered:
```/tmp/917414073/src/github.com/influxdata/influxdb/tsdb/index_test.go:490:6: tsResult declared but not used
couldn't load packages due to errors: github.com/influxdata/influxdb/tsdb_test
```